requirements.txt
# Core Dependencies
fastapi>=0.111.1,<0.112.0
uvicorn[standard]>=0.30.3,<0.31.0
sqlalchemy>=2.0.31,<2.1.0
pydantic[email]>=2.8.2,<2.9.0
python-dotenv>=1.0.1,<1.1.0
alembic>=1.13.2,<1.14.0
psycopg2-binary>=2.9.9,<2.10.0 # Driver para PostgreSQL (se for usar)
# asyncpg>=0.29.0,<0.30.0 # Driver async alternativo
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0 
pydantic_settings